requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==7.3.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
openai==1.60.1
phidata==2.7.9
primp==0.11.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
tomli==2.2.1
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
